pyproject.toml
[project]
name = "c12-treetypesmapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pygbif>=0.6.6",
    "python-dotenv>=1.2.1",
]
